refactor(anthropic-direct): extract the stream consumer and retry handler (3/7)
The coupled pair, done in one commit because they share the retry budget
and splitting them across commits would leave a broken contract in between.

  loop/stream-consumer.ts  drives translateMessageStream, yields translated
                           events, classifies how the round ended
  loop/round-retry.ts      trace phase + stream.retry reset + class-appropriate
                           backoff + post-backoff abort check
  loop/outcomes.ts         RoundOutcome / RetryOutcome discriminated unions

The consumer READS the budget (canRetryOverload / canRetryStreamIncomplete /
ttfbRetried); the retry handler SPENDS it. Both are handed the same
RoundRetryBudget instance built in runTurn — a copy would silently grant
every round an unlimited allowance, which is exactly the failure the stage 2
contract exists to prevent.

Control flow: sixteen bare `return`s and five `let` flags become values.
runTurn's round body is now a switch over RoundOutcome. `yield*` delegation
means event ORDER is preserved by construction, which is what makes the 82
existing end-to-end tests a real proof rather than a smoke check.

Ordering preserved deliberately:
  - retry.reset() still runs AFTER the retry branches and BEFORE the
    overload-exhausted / translator-errored terminals
  - overload-exhausted is still classified before translator-errored, so the
    clean terminal keeps winning over the fatal lane (#762)
  - ttfb/stall dispose() calls stay hand-duplicated at each exit path rather
    than consolidated into a finally; the handles are idempotent and the
    pre-split loop did exactly this

Both connection-phase and mid-stream TTFB re-drives now route through the one
handler, so the once-per-round allowance cannot drift between them.

Gates: pnpm lint clean; 526 tests pass across 51 files, unchanged.

Refs #360
